RF-AT_018.03.02 replace base methods to common BasePage; updated fill_in_field method in member_page.py
